[BE] fix: follow toId 수정

See merge request s08-bigdata-recom-sub2/S08P22C201!130

+++ back/tink/src/main/java/com/ssafy/tink/db/entity/FollowId.java
@@ -4,5 +4,5 @@ import java.io.Serializable;
 
 public class FollowId implements Serializable {
 	private int toId;
-	private int member;
+	private long member;
 }
